net/mlx5e: Use size_t to store byte offset in statistics descriptors
The byte offset of counter descriptors should be stored in size_t variable
instead of an integer.

+++ drivers/net/ethernet/mellanox/mlx5/core/en_stats.h
@@ -47,7 +47,7 @@
 
 struct counter_desc {
 	char		format[ETH_GSTRING_LEN];
-	int		offset; /* Byte offset */
+	size_t		offset; /* Byte offset */
 };
 
 struct mlx5e_sw_stats {
